[feat]:调整代码结构，分离耦合变量 [解决方案]:角色祈愿抽象为类CharacterGachaHelper [测试范围]:角色祈愿

+++ lib/random_gen.cpp
@@ -8,5 +8,5 @@ unsigned int seed_r = seed_gen() % 10000000;
 unsigned long long int seed_t =
     static_cast<unsigned long long int>(std::chrono::high_resolution_clock::now().time_since_epoch().count()) *
     static_cast<unsigned long long int>(100000);
-size_t seed = static_cast<size_t>(seed_t) + static_cast<size_t>(seed_r);
+int seed = static_cast<int>(seed_t) + static_cast<int>(seed_r);
 std::mt19937_64 randomgen::generatorz(seed);
